jormungandr: elasticsearch kraken for places api
